+++ src/3DViewer/mainwindow.h
@@ -44,13 +44,13 @@ class MainWindow : public QMainWindow {
   void on_dothColorButton_clicked();
   void on_backColorButton_clicked();
 
-  void on_thicknessBox_valueChanged(double arg1);
-  void on_sizeBox_valueChanged(double arg1);
+  void on_thicknessBox_valueChanged(int arg1);
+  void on_sizeBox_valueChanged(int arg1);
 
   void on_typeLine_currentIndexChanged(int index);
   void on_typeDoth_currentIndexChanged(int index);
 
-private:
+ private:
   Ui::MainWindow *ui;
   QTimer *timer;
   MyGLWidget *myGLWidget;
